\name{print.maboost}
\alias{print.maboost}

\title{
  Model Information for Ada
}

\description{
  \code{print} lists the model information and final confusion matrix
  for submitted data.
}

\usage{
\method{print}{maboost}(x, \dots)
}

\arguments{
  \item{x}{object generated by the function \code{maboost}. }
  \item{\dots}{other arguments not used by this function.}
}

\details{
  \code{print} produces a summary of the original function call, method
  used for boosting, number of iterations, final confusion matrix,
  error from data used to build the model, and estimates of M.
  
  Note:  any object of class \code{maboost} invokes \code{print}, when
  printed to the screen.
}

\value{
  No value returned.
}

\seealso{
  \code{\link{summary.maboost}},\code{\link{predict.maboost}},
 }
\keyword{methods}