\name{summary.maboost}
\alias{summary.maboost}

\title{
  Summary of model fit for arbitrary data (test, validation, or training)
}

\description{
  \code{summary} lists the model information for fitted model and final 
  confusion matrix.
}

\usage{
\method{summary}{maboost}(object, n.iter=NULL, \dots)
}

\arguments{
  \item{object}{object generated by 'maboost'.}
  \item{n.iter}{specific iteration to obtain the training and testing information at.}
  \item{\dots}{other arguments not used by this function.}
}

\details{
  \code{summary} produces a summary of the original function call, method
  used for boosting for a specific iteration,  accuracy, and kappa 
  statistic (a measure of agreement between the observed classification and 
  predicted classification) for the training data.  
  
  In addition, if any other data set (i.e. test or validation)
  has been incorporated to the \code{maboost} object, 
  \code{summary} produces analogous information.
}

\seealso{
  \code{\link{maboost}},\code{\link{predict.maboost}} 
}
\keyword{methods}