\name{update.maboost}
\alias{update.maboost}

\title{
 Add more trees to an maboost object
}

\description{
  \code{maboost.update} updates the \code{maboost} object to have additional trees given a new number of
   iterations.
}

\usage{
\method{update}{maboost}(object, x, y, test.x, test.y = NULL, n.iter, \dots)
}

\arguments{
  \item{object}{object generated by the function \code{maboost}. }
  \item{x}{x training data}
  \item{y}{training response}
  \item{test.x}{x testing data (optional)}
  \item{test.y}{the true labeling for this testing data (optional)}
  \item{n.iter}{ new number of iterations, must be provided and n.iter>iter}
  \item{\dots}{other arguments not used by this function.}
}

\value{
   updated \code{maboost} object.
}

\seealso{
  \code{\link{maboost}},\code{\link{summary.maboost}},\code{\link{predict.maboost}},
  }
\keyword{methods}