\name{varplot.maboost}
\alias{varplot.maboost}

\title{
  Variable selection with maboost
}

\description{
  \code{varplot.maboost} ranks the variable according to their importance.
}

\usage{

 varplot.maboost(x,plot.it=TRUE, type=c("none","scores"),max.var.show=30,\dots)
}

\arguments{
  \item{x}{object generated by the function \code{maboost}. }
  \item{plot.it}{flag to plot a figure}
   \item{type}{default="none", if scores is selected the scores of variables are returned. }
   \item{max.var.show}{default="30", set it to larger value if you have more than 30 variables. }
   \item{\dots}{other arguments not used by this function.}
}

\details{
  \code{varplot.maboost} This command provides a sense of variable importance. The more frequently a variable is selected for boosting, the more likely the variable contains useful information for classification. for rpart, the standard criteria explained in (Hastie et al,2001, pg332) is used while for C5.0. the function C5imp is used to calculate the variables importance for each C5.0 model and these scores are averaged over all trees.
}

\value{
 scores are returned.
}

\seealso{
  \code{\link{summary.maboost}},\code{\link{predict.maboost}},
 }
\keyword{methods}
